separate tasks into dedicated component
